requirements.txt
fastapi
uvicorn
sqlalchemy
databases
pydantic
aiobotocore
boto3
alembic
pytest
requests
asyncpg
